Add function to stop an event loop, along with socket cleanup code.
When we exit out of the event loop, we just close all the file
descriptors we know about.  Later, we should have a function to register
a callback, which could send any sort of disconnection messages before
closing.

Also, use pselect instead of select.  We also block SIGINT before
checking the flag, and pselect unblocks it.  This eliminates a race
condition when checking for the stop_loop flag; if a signal hits after
we check the flag but before we enter the select syscall.