Make check_mem abort if a malloc fails.
I don't really want to try to recover from a failed allocation right
now.  If a device is out of memory, lots of bad stuff will probably
happen if we try to hobble along.  It might be useful in the future
though, if we are operating on especially constrained embedded devices.