Rename "lib" directory to "deps" to follow GitHub conventions.
I think the "deps" directory is not scanned when generating GitHub
graphs and other project statistics (before, GitHub was saying that XMP3
was 61% shell script!).  Hopefully, this will fix it.

I found this info at [1], which is the list of directories ignored by
Linguist, the tool GitHub uses to parse code.

[1]: https://github.com/github/linguist/blob/master/lib/linguist/vendor.yml